refactor(member): static 값들을 mock/data로 이동(#274)
