--- apps/member/src/mocks/data/qnaBoardList.json
@@ -1,42 +1,0 @@
-[
-  {
-    "id": 1,
-    "type": "질문",
-    "title": "이 코드 오류 찾아주실 분? ㅜㅜ",
-    "writer": "김정은",
-    "createAt": "2021-09-01 00:00:00",
-    "contents": "안녕하세요"
-  },
-  {
-    "id": 2,
-    "type": "질문",
-    "title": "인텔리제이랑 이클립스 중에서 어떤 걸 써야 좋을까용",
-    "writer": "김정은",
-    "createAt": "2021-09-01 00:00:00",
-    "contents": "안녕하세요"
-  },
-  {
-    "id": 3,
-    "type": "질문",
-    "title": "노트북 바꾸려고 하는데 사양 한 번만 봐주세요...",
-    "writer": "김정은",
-    "createAt": "2021-09-01 00:00:00",
-    "contents": "안녕하세요"
-  },
-  {
-    "id": 4,
-    "type": "질문",
-    "title": "백엔드 공부 어떻게 시작해야 할까요",
-    "writer": "김정은",
-    "createAt": "2021-09-01 00:00:00",
-    "contents": "안녕하세요"
-  },
-  {
-    "id": 5,
-    "type": "질문",
-    "title": "내년도에 해커톤 있으면 같이 나갈 팀을 꾸리고 싶은데 혹시 있을까요",
-    "writer": "김정은",
-    "createAt": "2021-09-01 00:00:00",
-    "contents": "안녕하세요"
-  }
-]
